requirements.txt
altgraph==0.17.2
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
comtypes==1.1.10
Django==4.0.3
future==0.18.2
idna==3.3
MouseInfo==0.1.3
mysql-connector-python==8.0.28
pefile==2021.9.3
pocketsphinx @ file:///C:/Users/Matheus/Documents/python/pocketsphinx-0.1.15-cp310-cp310-win_amd64.whl
PyAudio @ file:///C:/Users/Matheus/Documents/python/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyInstaller @ file:///C:/Users/Matheus/Documents/python/PyInstaller-3.6-py2.py3-none-any.whl
PyMsgBox==1.0.9
pynput==1.7.5
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pyttsx3 @ file:///C:/Users/Matheus/Documents/python/pyttsx3-2.90-py3-none-any.whl
pytweening==1.0.4
pywin32==302
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
speech-recognition-fork @ file:///C:/Users/Matheus/Documents/python/speech_recognition_fork-3.8.1.2021.6.14-py2.py3-none-any.whl
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
vosk==0.3.32
